--- a/CHANGELOG.md
+++ b/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Optimized the portfolio calculations with smarter cloning of activities
 - Upgraded `webpack-bundle-analyzer` from version `4.10.1` to `4.10.2`
 
 ## 2.110.0 - 2024-09-24