--- a/apps/api/src/app/portfolio/calculator/twr/portfolio-calculator.ts
+++ b/apps/api/src/app/portfolio/calculator/twr/portfolio-calculator.ts
@@ -180,10 +180,10 @@ export class TWRPortfolioCalculator extends PortfolioCalculator {
     let valueAtStartDateWithCurrencyEffect: Big;
 
     // Clone orders to keep the original values in this.orders
-    let orders: PortfolioOrderItem[] = cloneDeep(this.activities).filter(
-      ({ SymbolProfile }) => {
+    let orders: PortfolioOrderItem[] = cloneDeep(
+      this.activities.filter(({ SymbolProfile }) => {
         return SymbolProfile.symbol === symbol;
-      }
+      })
     );
 
     if (orders.length <= 0) {